Moving Utility Classes to Util Package
