+++ src/main/java/io/servertap/api/v1/ServerApi.java
@@ -6,8 +6,8 @@ import io.javalin.http.NotFoundResponse;
 import io.javalin.http.ServiceUnavailableResponse;
 import io.javalin.openapi.*;
 import io.servertap.Constants;
-import io.servertap.Lag;
-import io.servertap.ServerExecCommandSender;
+import io.servertap.utils.Lag;
+import io.servertap.utils.ServerExecCommandSender;
 import io.servertap.api.v1.models.*;
 import io.servertap.mojang.api.MojangApiService;
 import io.servertap.mojang.api.models.NameChange;
