+++ src/main/java/io/servertap/utils/Lag.java
@@ -1,4 +1,4 @@
-package io.servertap;
+package io.servertap.utils;
 
 public class Lag implements Runnable {
